requirements.txt
langchain==0.1.16
langchain-community==0.0.32
langchain-openai==0.1.3
openai==1.17.1
python-decouple
loguru

# Tutorial 3
transformers==4.35.2
torch==2.1.0
einops==0.7.0
accelerate==0.26.1